pyproject.toml
[project]
name = "agent-radio"
version = "0.1.0"
description = "AI-powered radio stream with TTS announcements"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135",
    "uvicorn>=0.43",
    "pyyaml>=6.0",
    "kokoro>=0.9",
    "soundfile>=0.13",
    "numpy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
